third_party/blink/web_tests/platform/generic/fast/dom/Window/Location/ancestor-origins-expected.txt

ancestorOrigins.length = 0 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
ancestorOrigins.length = 1
ancestorOrigins[0] = file://
